set default value for jwt algorithm
