+++ app/utils/jwt.py
@@ -1,11 +1,10 @@
 import jwt
 from jwt import PyJWTError
 from datetime import datetime, timedelta
-from fastapi import Request, status, HTTPException
 from app.providers.config import configs
 
 class JWTUtils:
-    def __init__(self, secret_key, algorithm):
+    def __init__(self, secret_key, algorithm="HS256"):
         self.SECRET_KEY = secret_key
         self.ALGORITHM = algorithm
 
@@ -24,24 +23,4 @@ class JWTUtils:
             return None
 
 
-jwt_utils = JWTUtils(configs.secret_key, "HS256")
 
-async def verify_token(request: Request):
-    if configs.auth_enabled:
-        token = request.cookies.get("auth_token")
-        if not token:
-            raise HTTPException(
-                status_code=status.HTTP_401_UNAUTHORIZED,
-                detail="Authentication required"
-            )
-
-        payload = jwt_utils.decode_jwt_token(token)
-        if payload is None or "sub" not in payload:
-            raise HTTPException(
-                status_code=status.HTTP_401_UNAUTHORIZED,
-                detail="Invalid token"
-            )
-
-        return payload["sub"]
-    else:
-        return None
